requirements.txt
# Core FastAPI dependencies
fastapi==0.115.0
uvicorn[standard]==0.32.0
jinja2==3.1.4
python-multipart==0.0.22

# Supabase (Auth + Postgres)
supabase==2.28.0

# Environment & Configuration
python-dotenv==1.2.1

# Data validation
pydantic==2.12.5

rich==14.3.2
simple-websocket==1.1.0
six==1.17.0
SQLAlchemy==2.0.46
sqlmodel==0.0.32
starlette==0.52.1
storage3==2.28.0
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.28.0
supabase-auth==2.28.0
supabase-functions==2.28.0
tenacity==9.1.4
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
watchfiles==1.1.1
websockets==15.0.1
wrapt==2.1.1
wsproto==1.3.2
yarl==1.22.0
zstandard==0.25.0
plotly>=5.18.0
